requirements.txt
pinecone
langchain
langchain_community
langchain_huggingface
langchain_openai
langchain_pinecone
flask
pypdf
python-dotenv

-e .